Adjustments and refactor charts for reuse
