+++ components/analytics/index.tsx
@@ -1,3 +1,7 @@
 export { default as DefaultChartHeader } from "./DefaultChartHeader";
 export { default as DurationFilter } from "./DurationFilter";
 export { default as MovingAverageFilter } from "./MovingAverageFilter";
+export { default as APYChart } from "./APYChart";
+export { default as ProtocolChart } from "./ProtocolChart";
+export { default as TotalSupplyChart } from "./TotalSupplyChart";
+export { default as CollateralPieChart } from "./CollateralPieChart";
