+++ pages/analytics/collateral.tsx
@@ -3,9 +3,8 @@ import Head from "next/head";
 import Link from "next/link";
 import { Typography } from "@originprotocol/origin-storybook";
 import { GetServerSideProps } from "next";
-import classnames from "classnames";
 import { map } from "lodash";
-import { orderBy } from "lodash";
+import { find, orderBy } from "lodash";
 import {
   BarElement,
   CategoryScale,
@@ -21,6 +20,7 @@ import {
 import { aggregateCollateral, backingTokens } from "../../utils/analytics";
 import { fetchAllocation, fetchCollateral } from "../../utils/api";
 import { formatCurrency } from "../../utils/math";
+import { strategyMapping } from "../../constants";
 
 ChartJS.register(CategoryScale, LinearScale, BarElement);
 
@@ -61,52 +61,59 @@ const CollateralPoolDistributions = ({ data = [] }) => {
     <div className="flex flex-col space-y-6">
       <Typography.Body3>Collateral Distribution</Typography.Body3>
       <div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 3xl:grid-cols-4 gap-1">
-        {data?.map(({ name, address, holdings }) => (
-          <LayoutBox key={name}>
-            <div className="flex flex-col w-full h-full p-6">
-              <Typography.Body3>{name}</Typography.Body3>
-              <Typography.Caption2 className="text-subheading">
-                Collateral
-              </Typography.Caption2>
-              <div className="flex flex-wrap w-full gap-4 py-4">
-                {map(holdings, (holdingTotal, token) =>
-                  backingTokens[token] ? (
-                    <div
-                      key={token}
-                      className="flex flex-row space-x-3 items-center h-[40px] w-[135px]"
-                    >
-                      <Image
-                        src={backingTokens[token]?.logoSrc}
-                        width={28}
-                        height={28}
-                        alt={token}
-                      />
-                      <div className="flex flex-col">
-                        <Typography.Caption>{token}</Typography.Caption>
-                        <Link
-                          href={`https://etherscan.io/address/${address}`}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="flex flex-row space-x-1 items-center w-full"
-                        >
-                          <Typography.Caption className="text-subheading">
-                            {`${formatCurrency(holdingTotal, 2)}`}
-                          </Typography.Caption>
-                          <Image
-                            src="/images/link.svg"
-                            width={12}
-                            height={12}
-                            alt="External link"
-                          />
-                        </Link>
+        {data?.map(({ name, address, holdings }) => {
+          const strategyName =
+            find(
+              strategyMapping,
+              (item) => item.address?.toLowerCase() === address?.toLowerCase()
+            )?.short_name || name;
+          return (
+            <LayoutBox key={address}>
+              <div className="flex flex-col w-full h-full p-6">
+                <Typography.Body3>{strategyName}</Typography.Body3>
+                <Typography.Caption2 className="text-subheading">
+                  Collateral
+                </Typography.Caption2>
+                <div className="flex flex-wrap w-full gap-4 py-4">
+                  {map(holdings, (holdingTotal, token) =>
+                    backingTokens[token] ? (
+                      <div
+                        key={token}
+                        className="flex flex-row space-x-3 items-center h-[40px] w-[135px]"
+                      >
+                        <Image
+                          src={backingTokens[token]?.logoSrc}
+                          width={28}
+                          height={28}
+                          alt={token}
+                        />
+                        <div className="flex flex-col">
+                          <Typography.Caption>{token}</Typography.Caption>
+                          <Link
+                            href={`https://etherscan.io/address/${address}`}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="flex flex-row space-x-1 items-center w-full"
+                          >
+                            <Typography.Caption className="text-subheading">
+                              {`${formatCurrency(holdingTotal, 2)}`}
+                            </Typography.Caption>
+                            <Image
+                              src="/images/link.svg"
+                              width={12}
+                              height={12}
+                              alt="External link"
+                            />
+                          </Link>
+                        </div>
                       </div>
-                    </div>
-                  ) : null
-                )}
+                    ) : null
+                  )}
+                </div>
               </div>
-            </div>
-          </LayoutBox>
-        ))}
+            </LayoutBox>
+          );
+        })}
       </div>
     </div>
   );
